Use a constant userID.
The reason this wasn't working is because when you create a new
Reflect instance it forks from the latest confirmed state from
server. It doesn't fork pending local changes.

The flow you had was:
- create new reflect A
- mutate A
- create new reflect B

The mutation to A was always pending when you created B, thus
it was not in B.

An alternate solution would be to track when the mutations were
confirmed by server (by having server write something into state
that client can observe) and delay creating B until that time.